Update README.md
Developing-Data-Products-R-Markdown-Presentation-Plotly This is a repo with the gh-pages branch. The HTML page can be viewed online by clicking on this Github Pages link. R markdown and compiled HTML (Source code) files can be found in the repository itself. Synopsis Following instructions have been given for the assignment -

Create a web page presentation using R Markdown that features a plot created with Plotly. Host your webpage on either RPubs, GitHub Pages, or NeoCities. Your webpage must contain the date that you created the document, and it must contain a plot created with Plotly. The Interactive Plots presented in this Assignment are as follows -

A Heatmap depicting the Daily Ozone Levels in New York over a period of 5 months (May to September 1973). A Time-Series chart of the Population of the United States (in millions) for the period 1790-1970.